# swagger.yaml
openapi: 3.0.3
info:
  version: 1.0.1
  title: MediRecord
  description: |-
    It will be a parsonal medical application, This app targeted for every person of the society either diseased or not. The Primary function of this app is to track and maintain a persons all medical and this related data. This application should be able to contain a persons general information, physical health status, data of each time met with consultant including consultants info, medical tests and Their results,  medication and routine tests information. There should be all the advises of the consultants.
servers: 
  - description: Local server
    url: "http://localhost:8000"
...
paths:
...
  "/general-condition/{id}":
    get:
      security:
        - bearerAuth: []    
      tags: [General Condition]
      description: Get a General Condition of a patient"
      parameters: 
        - $ref: "#/components/parameters/idPathParam"
        - in: query
          name: expand
          description: expand Prescription and Medicines
          schema:
            type: string
            example: "prescription, medicine" 
      responses:
        "200":
          description: "Return a condition"
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/General-Condition"
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: url
                        example: "/general-condition/1"
                        
        "400":
          $ref: "#/components/responses/400"
        "500":
          "$ref": "#/components/responses/500"    
    patch:       
      security:
        - bearerAuth: []  
      tags: [General Condition]
      description: Partially update an Condition
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      requestBody:
        content:
          "application/json":
            schema:
              type: object
              properties:
                sign_symptom:
                  type: string
                  example: fever
                consultant_visit_id:
                  type: integer
                  format: int32
                  example: 1
                BP:
                  type: integer
                  format: int32
                  example: 80-90
                pulse:
                  type: integer
                  format: int32
                  example: 62
                weight:
                  type: integer
                  format: int32
                  example: 62
                age:
                  type: integer
                  format: int32
                  example: 32  
                body_condition:
                  type: string 
                  example: bad  
      responses:
        "200":
          description: "Successfully Updated"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/GeneralUpdateDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
    delete:
      security:
        - bearerAuth: []
      tags: [General Condition]
      description: Delete a Condition by id
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      responses:
        "204":
          description: "Condition deleted successfully"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"          
...
  "/medicine/{id}":
    get:
      security:
        - bearerAuth: []    
      tags: [Medicine]
      description: Get a medicine of a patient"
      parameters: 
        - $ref: "#/components/parameters/idPathParam"
        - in: query
          name: expand
          description: expand Prescription 
          schema:
            type: string
            example: "prescription" 
      responses:
        "200":
          description: "Return a medicine"
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Medicine"
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: url
                        example: "/medicine/1"
                        
        "400":
          $ref: "#/components/responses/400"
        "500":
          "$ref": "#/components/responses/500"    
    patch:       
      security:
        - bearerAuth: []  
      tags: [Medicine]
      description: Partially update a medicine
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      requestBody:
        content:
          "application/json":
            schema:
              type: object
              properties:
                id:
                  type: string
                  example: 1 
                prescription:
                  type: string
                  example: 1
                dose:
                  type: string 
                  example: "2 times in a a"
                duration:
                  type: string 
                  example: '9 day' 
      responses:
        "200":
          description: "Successfully Updated"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/MedicineUpdateDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
    delete:
      security:
        - bearerAuth: []
      tags: [Medicine]
      description: Delete a medicine by id
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      responses:
        "204":
          description: "Medicine deleted successfully"
        "401":
          $ref: "#/components/responses/401"
        "404": 
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"   
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Timestamp:
      type: object
      properties:
        createdAt:
          type: string
          format: date
        updatedAt:
          type: string
          format: date
      required: [createdAt, updatedAt]
...
    General-Condition:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int32
              example: 1
            sign_symptom:
              type: string 
              example: fever
            consultant_visit_id:
              type: string
              example: 1
            BP:
              type: string
              example: 80-90
            pulse:
              type: integer
              format: int32
              example: 62
            weight:
              type: integer
              format: int32
              example: 62
            age:
              type: integer
              format: int32
              example: 32  
            body_condition:
              type: string 
              example: bad  
          required: [id,consultant_visit_id,BP,pulse,weight,age,body_condition]
        - $ref: "#/components/schemas/Timestamp"
...
    GeneralUpdateDTO:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 200
        message:
          type: string
          example: General Condition updated successfully
        data:
          $ref: "#/components/schemas/General-Condition"
        links:
          type: object
          properties:
            self:
              type: string
              example: "/general-condition/1"              
...
    Medicine:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int32
              example: 1 
            prescription_id:
              type: integer
              format: int32
              example: 1
            dose:
              type: string 
              example: "2 times in a a"
            duration:
              type: string 
              example: '9 day' 
          required: [prescription_id,dose,duration ]
        - $ref: "#/components/schemas/Timestamp"
...
    MedicineUpdateDTO:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 200
        message:
          type: string
          example: Prescription Updated successfully"
        data:
          $ref: "#/components/schemas/Medicine"
        links:
          type: object
          properties:
            self:
              type: string
              format: url
              example: "/medicine"           
...
  responses:
    '400':
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 400
              error:
                type: string
                example: Bad Request
              data:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: email
                    message:
                      type: string
                      example: Invalid email
                    in:
                      type: string
                      enum:
                        - query
                        - body
                      example: body
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 401
              error:
                type: string
                example: Unauthorized
              message:
                type: string
                example: You don't have the right permissions
    '404':
      description: Resource Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 404
              error:
                type: string
                example: Not Found
              message:
                type: string
                example: Requested resource not found
    '500':
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 500
              error:
                type: string
                example: Internal Server Error
              message:
                type: string
                example: We are sorry. Please try again later.
  parameters:
...
    idPathParam:
      in: path
      name: id
      required: true
      description: Prescription ID
      schema:
        type: string
        example: 1